app/api/upload.py
from fastapi import APIRouter, UploadFile, File, Depends, HTTPException
from app.core import s3
from fastapi.security import OAuth2PasswordBearer
from app.database import get_db
from sqlalchemy.orm import Session
from app.models import models
from jose import jwt, JWTError
from app.core.security import SECRET_KEY, ALGORITHM

router = APIRouter()
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/login")

def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        email = payload.get("sub")
        user = db.query(models.User).filter(models.User.email == email).first()
        if not user:
            raise HTTPException(status_code=401, detail="User not found")
        return user
    except JWTError:
        raise HTTPException(status_code=401, detail="Invalid token")

@router.post("/")
def upload_file(file: UploadFile = File(...), current_user: models.User = Depends(get_current_user)):
    key = f"{current_user.id}/{file.filename}"
    url = s3.upload_to_s3(file, key)
    return {"file_url": url}
